- Добавить возможность редактировать название решения

+++ src/main/java/com/example/servingwebcontent/repositories/QuizDecisionRepository.java
@@ -1,11 +1,11 @@
 package com.example.servingwebcontent.repositories;
 
-import com.example.servingwebcontent.domain.quiz.QuizDecision;
+import com.example.servingwebcontent.domain.quiz.decision.QuizDecision;
 import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface QuizDecisionRepository extends CrudRepository<QuizDecision, Integer> {
+public interface QuizDecisionRepository extends CrudRepository<QuizDecision, Long> {
     QuizDecision findByName(String name);
 
     List<QuizDecision> findAllByOrderByName();
