+++ src/main/java/com/example/servingwebcontent/repositories/QuizRepository.java
@@ -5,7 +5,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import java.util.List;
 
-public interface QuizRepository extends CrudRepository<Quiz, Integer> {
+public interface QuizRepository extends CrudRepository<Quiz, Long> {
 
     Quiz findByName(String name);
     Quiz findByShortName(String quizShortName);
