+++ src/main/java/com/example/servingwebcontent/repositories/QuizTaskRepository.java
@@ -4,6 +4,6 @@ import com.example.servingwebcontent.domain.quiz.Quiz;
 import com.example.servingwebcontent.domain.quiz.QuizTask;
 import org.springframework.data.repository.CrudRepository;
 
-public interface QuizTaskRepository extends CrudRepository<QuizTask, Integer> {
+public interface QuizTaskRepository extends CrudRepository<QuizTask, Long> {
     QuizTask findByPositionAndQuiz(int pos, Quiz quiz);
 }
